rename: move a MainActivity and separate a composable function
