+++ app/src/main/java/com/jukco/waitforme/MainScreen.kt
@@ -1,14 +1,8 @@
-package com.jukco.waitforme.ui
+package com.jukco.waitforme
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
@@ -16,29 +10,11 @@ import androidx.navigation.compose.rememberNavController
 import com.jukco.waitforme.ui.navi.BottomNaviBar
 import com.jukco.waitforme.ui.navi.BottomNaviItem
 import com.jukco.waitforme.ui.navi.NavigationGraph
-import com.jukco.waitforme.ui.theme.WaitForMeTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            WaitForMeTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    MainView()
-                }
-            }
-        }
-    }
-}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Preview(showSystemUi = true)
 @Composable
-fun MainView() {
+fun MainScreen() {
     val navController = rememberNavController()
 
     Scaffold (
